+++ src/components/planShareRoom/sideBar/placeSearchArea/placeSearchAreaContent/PlaceSearchAreaContent.tsx
@@ -1,0 +1,11 @@
+import PlaceSearchLists from "./PlaceSearchLists";
+
+const PlaceSearchAreaContent = () => {
+  return (
+    <div className="bg-white">
+      <PlaceSearchLists />
+    </div>
+  );
+};
+
+export default PlaceSearchAreaContent;
