+++ src/components/planShareRoom/sideBar/placeSearchArea/placeSearchAreaHeader/PlaceSearchCategory.tsx
